requirements.txt
flask>=2.0.0,<3.0.0
flask-cors>=3.0.0,<5.0.0
mysql-connector-python>=8.0.0,<9.0.0
fpdf>=1.7.2,<3.0.0
flask-wtf>=1.0.0,<2.0.0
werkzeug>=2.0.0,<3.0.0
flask-limiter>=3.0.0,<4.0.0
bleach>=6.0.0,<7.0.0
flask-login>=0.6.3,<1.0.0
